refactor(was): business package -> app package로 이름 변경
