+++ src/main/java/app/Application.java
@@ -1,0 +1,10 @@
+package app;
+
+import app.config.AppConfig;
+import infra.webserver.WebServer;
+
+public class Application {
+    public static void main(String[] args) throws Exception {
+        WebServer.run(AppConfig.getInstance().getControllerConfig().getFrontController(), args);
+    }
+}
