+++ src/main/java/app/home/controller/HomeControllerConstants.java
@@ -1,0 +1,8 @@
+package app.home.controller;
+
+public final class HomeControllerConstants {
+
+    private HomeControllerConstants(){}
+
+    public static final byte[] HELLO_WORLD_BYTES = "Hello world".getBytes();
+}
