+++ src/main/java/app/user/dao/UserDaoImpl.java
@@ -1,0 +1,24 @@
+package app.user.dao;
+
+import app.user.User;
+import app.user.repository.UserRepository;
+import java.util.List;
+
+public class UserDaoImpl implements UserDao {
+
+    private final UserRepository userRepository;
+
+    public UserDaoImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public User save(User user) {
+        return userRepository.save(user);
+    }
+
+    @Override
+    public List<User> findAll() {
+        return userRepository.findAll();
+    }
+}
