+++ src/main/java/app/user/db/UserDatabase.java
@@ -1,0 +1,8 @@
+package app.user.db;
+
+import auth.AuthUserDetails;
+import auth.db.AuthLoginDatabase;
+import infra.db.Database;
+
+public interface UserDatabase extends Database<AuthUserDetails>, AuthLoginDatabase {
+}
