--- src/main/java/businuess/home/controller/HomeControllerConstants.java
@@ -1,8 +1,0 @@
-package businuess.home.controller;
-
-public final class HomeControllerConstants {
-
-    private HomeControllerConstants(){}
-
-    public static final byte[] HELLO_WORLD_BYTES = "Hello world".getBytes();
-}
