--- src/main/java/businuess/user/User.java
@@ -1,97 +1,0 @@
-package businuess.user;
-
-import auth.AuthUserDetails;
-import java.util.Objects;
-
-public class User implements AuthUserDetails {
-    private final String userId;
-    private final String password;
-    private final String name;
-    private final String email;
-
-    @Override
-    public String getUserId() {
-        return userId;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String getEmail() {
-        return email;
-    }
-
-    public static UserBuilder builder() {
-        return new UserBuilder();
-    }
-
-    private User(String userId, String password, String name, String email) {
-        this.userId = userId;
-        this.password = password;
-        this.name = name;
-        this.email = email;
-    }
-
-    @Override
-    public String toString() {
-        return "User [userId=" + userId + ", password=" + password + ", name=" + name + ", email=" + email + "]";
-    }
-
-    public static class UserBuilder {
-
-        private String userId;
-        private String password;
-        private String name;
-        private String email;
-
-        public UserBuilder userId(String userId) {
-            this.userId = userId;
-            return this;
-        }
-
-        public UserBuilder password(String password) {
-            this.password = password;
-            return this;
-        }
-
-        public UserBuilder name(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public UserBuilder email(String email) {
-            this.email = email;
-            return this;
-        }
-
-        public User build() {
-            return new User(userId, password, name, email);
-        }
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        User user = (User) o;
-        return Objects.equals(userId, user.userId) && Objects.equals(password, user.password)
-                && Objects.equals(name, user.name) && Objects.equals(email, user.email);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(userId, password, name, email);
-    }
-}
