--- src/main/java/businuess/user/dao/UserDaoImpl.java
@@ -1,24 +1,0 @@
-package businuess.user.dao;
-
-import businuess.user.User;
-import businuess.user.repository.UserRepository;
-import java.util.List;
-
-public class UserDaoImpl implements UserDao {
-
-    private final UserRepository userRepository;
-
-    public UserDaoImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-
-    @Override
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-}
