--- src/main/java/businuess/user/db/UserDatabase.java
@@ -1,8 +1,0 @@
-package businuess.user.db;
-
-import auth.AuthUserDetails;
-import auth.db.AuthLoginDatabase;
-import infra.db.Database;
-
-public interface UserDatabase extends Database<AuthUserDetails>, AuthLoginDatabase {
-}
